Clear newly-built ships' part transport.
A ship may be destroyed and rebuilt the same turn.  If it was carrying
components, we must reset its cargo.  Unfortunately, there is no easy way to
detect that.

The approach I'm taking here is to scan UTIL.TMP fo "ship built" records,
and reset all ships seen.

An alternative approach would be to hook into a pcontrol stage shortly after
combat, which requires extra setup for hosts.